requirements.txt
matplotlib
numpy
pandas
Pillow
tensorboard
PyYAML
torch
